requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap4==2.3.1
django-cors-headers==3.5.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
djangorestframework==3.12.2
phonenumbers==8.12.13
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1
whitenoise==5.2.0
